+++ app/api.py
@@ -5,10 +5,11 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from app.data import MongoDB
 from app.model import MatcherSortSearch
-from app.routers import (record_router,
-                         collection_router,
-                         graph_router,
-                         model_router)
+from app.routers import (mentor_router,
+                         mentee_router,
+                         data_router,
+                         model_router,
+                         graph_router)
 
 
 API = FastAPI(
@@ -16,10 +17,6 @@ API = FastAPI(
     version="0.49.1",
     docs_url='/',
 )
-
-@API.post("/info", tags=["General Operations"])
-async def info():
-    return {"info": "API Information"}
 
 API.db = MongoDB()
 API.matcher = MatcherSortSearch()
@@ -30,6 +27,10 @@ API.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
+
+@API.post("/info", tags=["General Operations"])
+async def info():
+    return {"info": "API Information"}
 
 
 @API.get("/version")
@@ -45,5 +46,5 @@ async def collections():
     return {"result": API.db.get_database_info()}
 
 
-for router in (record_router, collection_router, graph_router, model_router):
+for router in (mentor_router, mentee_router, data_router, model_router, graph_router):
     API.include_router(router.Router)
